http2: Add opt-in option to Framer to allow DataFrame struct reuse
The existing Framer in net/http2 allocates a new DataFrame struct
for each DataFrame read on calls to ReadFrame. On long-lived http2
streams that use lots of small data frames, these DataFrame structs
can be a significant source of memory allocation. The
SetReuseFrame option introduced here, if set on a Framer, allows the
Framer to reuse Frame objects and changes the ReadFrame API
so that returned Frame objects are only valid until the next call
to ReadFrame. This opt-in API now only implements reuse of DataFrames,
but it allows the Framer to reuse of any type of Frame object.

The footprint caused by creation of new DataFrame structs per data
frame was noticed in micro benchmarks of "gRPC" server "streaming
throuhgput", which uses the Framer in this package. After other
experimental changes to the benchmark that reduce allocations,
DataFrame structs appeared to a be significant source of allocations.
In this benchmark, 6,400 concurrent http2 streams on the server
did a total of roughly 900,000 round trips per second (each round
trip involves request and response http2 data frames with payloads of about
10 bytes each). Over a roughly 35 second benchmark period, this caused
slightly more than 30 million allocations of DataFrame structs, totalling
about 1.4 GB of allocated space. Reusing DataFrames in the Framer eliminated these
allocations and was found to improve the benchmark's throughput by about
5%.

Fixes golang/go#18502

Change-Id: Iad93420ef6c3918f54249d867098f1dadfa324d8